requirements/base.txt
Django==6.0.3
djangorestframework==3.17.1
python-decouple>=3.8
psycopg2-binary>=2.9
